+++ sdk/serialconsole/Azure.ResourceManager.SerialConsole/CHANGELOG.md
@@ -8,7 +8,7 @@ This is the first preview of the Azure Resource Manager SerialConsole client lib
 
 This package follows the [new Azure SDK guidelines](https://azure.github.io/azure-sdk/general_introduction.html), and provides many core capabilities:
 
-    - Support MSAL.NET, Azure.Identity is out of box for supporting MSAL.NET.
+    - Support MSAL.NET through Azure.Identity.
     - Support [OpenTelemetry](https://opentelemetry.io/) for distributed tracing.
     - HTTP pipeline with custom policies.
     - Better error-handling.